cody salsa

+++ src/main/java/com/project/bro_grammers/dto/CodeSubmissionRequest.java
@@ -9,6 +9,6 @@ import org.springframework.web.multipart.MultipartFile;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CodeSubmissionRequest {
-    private Integer uploaderId;
+    private Long uploaderId;
     private MultipartFile file;
 }
